fix: seperate App component
